Move `parseFees` to `portfolioSlice`
