+++ dcapal-frontend/src/components/allocationFlow/portfolioSlice.js
@@ -78,6 +78,46 @@ export const parseFeeType = (typeStr) => {
   return null;
 };
 
+export const parseFees = (fees) => {
+  if (!fees) return null;
+
+  const feeType = parseFeeType(fees.feeStructure.type);
+  if (!feeType) return null;
+
+  const parsed = getDefaultFees(feeType);
+
+  if (feeType === FeeType.ZERO_FEE) return parsed;
+
+  if (feeType === FeeType.FIXED) {
+    if (fees.maxFeeImpact) {
+      parsed.maxFeeImpact = fees.maxFeeImpact;
+    }
+    if (fees.feeStructure.feeAmount) {
+      parsed.feeStructure.feeAmount = fees.feeStructure.feeAmount;
+    }
+
+    return parsed;
+  }
+
+  if (feeType === FeeType.VARIABLE) {
+    if (fees.maxFeeImpact) {
+      parsed.maxFeeImpact = fees.maxFeeImpact;
+    }
+    if (fees.feeStructure.feeRate) {
+      parsed.feeStructure.feeRate = fees.feeStructure.feeRate;
+    }
+    if (fees.feeStructure.minFee) {
+      parsed.feeStructure.minFee = fees.feeStructure.minFee;
+    }
+    if (fees.feeStructure.maxFee) {
+      parsed.feeStructure.maxFee = fees.feeStructure.maxFee;
+    }
+    return parsed;
+  }
+
+  return null;
+};
+
 export const getDefaultFees = (type) => {
   if (!type) {
     return null;
